Add files via upload
This algorithm is created for Turkish Stock Market futures (BIST Viop)

What if we buy stocks that decreased in value for the last two consecutive days ?
Here's the answer with the backtest results. Looks like some stocks have reversal patterns.

The algorithm is basic and should not be used for trading purposes.